Add files via upload
updated 11-19 with nearly done figures. Some color adjustments need to be made, and I would like to reorder the genes for both my dot plots. I'm also saving this version because I'm going to try and run some stats and change my plot for Figure 7 and I'm afraid of ruining it entirely so I'm keeping this version here.